app/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Pillow==8.0.1
numpy==1.19.4
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.7.0+cpu 
torchvision==0.8.1+cpu